feat: rename + finish more refactors
