+++ frontend/app/src/pages/main/v1/workflow-runs-v1/$run/index.tsx
@@ -27,7 +27,7 @@ import WorkflowRunVisualizer from './v2components/workflow-run-visualizer-v2';
 import { useAtom } from 'jotai';
 import { preferredWorkflowRunViewAtom } from '@/lib/atoms';
 import { JobMiniMap } from './v2components/mini-map';
-import { useWorkflowDetails } from '../hooks/use-workflow-details';
+import { useWorkflowDetails } from '../hooks/workflow-details';
 
 export const WORKFLOW_RUN_TERMINAL_STATUSES = [
   WorkflowRunStatus.CANCELLED,
